fix: arruma arquivos e router
